HW8: add controllers and update routes
