+++ controllers/index.js
@@ -1,0 +1,9 @@
+import userControllers from './userControllers';
+import productControllers from './productControllers';
+import cityControllers from './cityControllers';
+
+export {
+    userControllers,
+    productControllers,
+    cityControllers
+}
